requirements.txt
qdrant-client
google-genai
python-dotenv
tiktoken
numpy